Add files via upload
Final, working version of game. This is intended to be used and modified to your liking for machine vision for your Doomsday robot. All the telemetry values could be used to make the robot do stuff via Arduino; I haven't finished mine yet, just tested that it WILL work with a couple of servos and some char triggers. Feel free to improve it, optimize it. The game takes place in the space in front of the robot. Make sure yours is safe to be around ie not spinning Battlebot blades or such. Mine uses Nerf Prometheus blasters and even then, they are disabled with a switch on the machine until I activate them personally. 
Includes the Processing data folder in 4 parts (font, jpegs, gifs and mp3s). The Arduino sketch is a start. Intended for Processing 2.x (I used 2.7) in 32-bit Java mode. Apparently the Serial class in Processing doesn't play well with 64 bit Java. 
If you get an arrayIndexOutOfBounds 0 exception when you run this, be sure you have already attached the Arduino to your PC USB or equivalent and have the Arduino sketch uploaded on it. 
Extract all the files, put the Arduino one in your Arduino sketchbook, it uses the elaspedMillis library found via Arduino. Put the Processing sketch in your Processing sketchbook. 
Note: the jpegs will have to be resized if your monitor is not 1366x768, or else the game won't launch. Plenty of free online converters can do this. 
Drivers etc for the Kinect not included here. This is intended for a Windows 7/8 machine, works well enough with mine 2GB RAM and some old decent enough processor I can't remember at the moment.